Fixes json response data, finishes services
When we're only asking for a single quote (since I changed the URL),
we only get back a single object in the quote part of the json response,
rather than an array.

Figured I'd swap it to just support that single response since that's
all we're really caring about right now.